requirements.txt
# 深度学习框架
torch
torchvision
torchaudio

# 数据处理
numpy
opencv-python
scipy
matplotlib
Pillow

# 3D处理
open3d

# 数据科学
pandas
scikit-learn

# 可视化
seaborn

# 工具库
tqdm
pyyaml
easydict

# 评估指标
pycocotools

# 其他工具
tensorboard 